+++ ext/torch/ext.cpp
@@ -2,8 +2,10 @@
 
 #include <rice/rice.hpp>
 
+void init_fft(Rice::Module& m);
 void init_linalg(Rice::Module& m);
 void init_nn(Rice::Module& m);
+void init_special(Rice::Module& m);
 void init_tensor(Rice::Module& m, Rice::Class& c, Rice::Class& rb_cTensorOptions);
 void init_torch(Rice::Module& m);
 
@@ -28,7 +30,9 @@ void Init_ext()
   init_torch(m);
   init_tensor(m, rb_cTensor, rb_cTensorOptions);
   init_nn(m);
+  init_fft(m);
   init_linalg(m);
+  init_special(m);
 
   init_cuda(m);
   init_device(m);