+++ test/test_helper.rb
@@ -11,7 +11,11 @@ class Minitest::Test
   def assert_elements_in_delta(expected, actual)
     assert_equal expected.size, actual.size
     expected.zip(actual) do |exp, act|
-      assert_in_delta exp, act
+      if exp.finite?
+        assert_in_delta exp, act
+      else
+        assert_equal exp, act
+      end
     end
   end
 end